move changes safely into students/[student_id]/goals[goal_id], delete temp_goals and goals pages folder
